+++ app/models/adjustments.py
@@ -10,13 +10,13 @@ class AdjustmentBase(BaseModel):
     distance_to_metro: float = Field(description="Корректировка на удаленность от метро")
     quality: float = Field(description="Корректировка на отделку")
 
-    price_trade: float = Field(description="Цена после корректировки на торг")
-    price_floor: float = Field(description="Цена после корректировки на торг")
-    price_area: float = Field(description="Цена после корректировки на площадь")
-    price_kitchen: float = Field(description="Цена после корректировки на площадь кухни")
-    price_balcony: float = Field(description="Цена после корректировки на наличие балкона")
-    price_metro: float = Field(description="Цена после корректировки на удаленность от метро")
-    price_final: float = Field(description="Цена после корректировки на отделку")
+    price_trade: int = Field(description="Цена после корректировки на торг")
+    price_floor: int = Field(description="Цена после корректировки на торг")
+    price_area: int = Field(description="Цена после корректировки на площадь")
+    price_kitchen: int = Field(description="Цена после корректировки на площадь кухни")
+    price_balcony: int = Field(description="Цена после корректировки на наличие балкона")
+    price_metro: int = Field(description="Цена после корректировки на удаленность от метро")
+    price_final: int = Field(description="Цена после корректировки на отделку")
 
 
 class AdjustmentGet(AdjustmentBase):
